move FleetStatsCCalculatorXXX from taxi do dvrp (to reuse in drt)
