+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/benchmark/TaxiBenchmarkStats.java
@@ -2,7 +2,7 @@ package org.matsim.contrib.taxi.benchmark;
 
 import org.apache.commons.math3.stat.descriptive.SummaryStatistics;
 import org.matsim.contrib.dvrp.data.Fleet;
-import org.matsim.contrib.taxi.util.stats.FleetStatsCalculator;
+import org.matsim.contrib.dvrp.fleet.FleetStatsCalculator;
 import org.matsim.contrib.taxi.util.stats.TaxiStats;
 import org.matsim.contrib.taxi.util.stats.TaxiStatsCalculator;
 import org.matsim.contrib.util.CSVLineBuilder;
